+++ ui/console_main.css
@@ -1,0 +1,13 @@
+#icon, #icon * {
+    width: 35px !important;
+    font-size: 24px !important;
+    line-height: 1.1 !important;
+    flex: none !important;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+#button, #button * {
+    width: 180px !important;
+    flex: none !important;
+}
